requirements.txt
line-bot-sdk
fastapi
uvicorn
typing
python-dotenv
pyquery